backend/jwt-lambda/requirements.txt
requests
PyJWT==2.6.0